modified post model
