+++ threddit_app/lib/features/community/view/create_community.dart
@@ -17,7 +17,6 @@ class _CreateCommunityState extends ConsumerState<CreateCommunity> {
   final TextEditingController _communityNameController =
       TextEditingController();
   String communityType = 'Public';
-  // ignore: unused_field
   CommunityType _type = CommunityType.Public;
   bool is18plus = false;
 
@@ -48,7 +47,6 @@ class _CreateCommunityState extends ConsumerState<CreateCommunity> {
       );
       return;
     }
-    //TODO try to fetch communities with similar name first
     final createCommunityFuture = ref.watch(createCommunityProvider(
         CreateCommunityParams(
             name: _communityNameController.text,
@@ -206,12 +204,7 @@ class _CreateCommunityState extends ConsumerState<CreateCommunity> {
                     child: Padding(
                       padding: const EdgeInsets.symmetric(horizontal: 20.0),
                       child: FilledButton(
-                        onPressed: (){
-                          Navigator.of(context).pop();
-                           Navigator.pushNamed(
-                            context, RouteClass.communityScreen);
-                            createCommunity();
-                        },
+                        onPressed: createCommunity,
                         style: const ButtonStyle(
                             backgroundColor:
                                 MaterialStatePropertyAll<Color>(Colors.blue)),
