+++ threddit_app/lib/features/home_page/view/screens/main_screen_layout.dart
@@ -1,12 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:threddit_clone/features/chatting/view/screens/main_chat_screen.dart';
+//import 'package:threddit_clone/features/chatting/view/screens/main_chat_screen.dart';
 import 'package:threddit_clone/features/post/view/add_post_screen.dart';
 import 'package:threddit_clone/features/home_page/view/screens/main_community_screen.dart';
 import 'package:threddit_clone/features/home_page/view/screens/home_screen.dart';
 import 'package:threddit_clone/features/home_page/view/screens/notifications_screen.dart';
 import 'package:threddit_clone/theme/colors.dart';
 import 'package:threddit_clone/features/home_page/view_model/home_page_provider.dart';
+import 'package:threddit_clone/features/community/view/community_screen.dart';
 
 class MainScreenLayout extends ConsumerStatefulWidget {
   const MainScreenLayout({super.key});
@@ -21,7 +22,7 @@ class _MainScreenLayout extends ConsumerState<MainScreenLayout> {
     const HomeScreen(),
     const MainCommunityScreen(),
     const AddPostScreen(),
-    const MainChatScreen(uid: 'user2'),
+    const CommunityScreen(id:"Sample Subreddit",uid:'user2'),                
     const NotificationsScreen()
   ];
