+++ threddit_app/lib/features/posting/model/repository/post_repository.dart
@@ -1,4 +1,4 @@
-import 'package:threddit_clone/models/post.dart';
+import 'package:threddit_clone/models/post101.dart';
 import 'dart:async';
 import 'dart:convert';
 import 'package:http/http.dart' as http;
