+++ threddit_app/lib/features/posting/view/screens/post_screen.dart
@@ -1,21 +1,17 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:threddit_clone/app/route.dart';
 import 'package:threddit_clone/features/commenting/view/widgets/comment_item.dart';
 import 'package:threddit_clone/features/commenting/view/widgets/add_comment.dart';
 
 import 'package:threddit_clone/features/home_page/view/widgets/right_drawer.dart';
-import 'package:threddit_clone/features/home_page/view/widgets/right_drawer_buttons.dart';
 import 'package:threddit_clone/features/reporting/view/report_bottom_sheet.dart';
 
-import 'package:threddit_clone/models/post.dart';
+import 'package:threddit_clone/models/post101.dart';
 import 'package:threddit_clone/features/posting/view/widgets/post_card.dart';
 import 'package:threddit_clone/theme/colors.dart';
 import 'package:threddit_clone/features/commenting/view_model/comment_provider.dart';
-import 'package:threddit_clone/theme/photos.dart';
-import 'package:threddit_clone/theme/text_styles.dart';
+
 
 class PostScreen extends ConsumerStatefulWidget {
   final Post currentPost;
